+++ docs/changelog.md
@@ -269,7 +269,7 @@
 - simplify methods impls: `take_ascii_ident`, `take_ascii_ident_tail`.
 
 ### text::str
-- new type `StrBuf`.
+- new types: `StrBuf`, `StringSmallAlloc`.
 - make `StringU*` use `NonMax*` for the length.
 - remove `StringU32`, `StringUSize`.